requirements.txt
attrs==23.1.0
azure-devops==6.0.0b2
beautifulsoup4==4.11.1
blessed==1.20.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==2.1.1
contourpy==1.1.1
cycler==0.12.0
execnet==2.0.2
fonttools==4.43.0
idna==3.4
importlib-resources==6.1.0
influxdb==5.3.1
iniconfig==2.0.0
inquirer==2.10.0
isodate==0.6.1
json2html==1.3.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.3
msgpack==1.0.7
msrest==0.6.21
numpy==1.24.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
py==1.11.0
pyparsing==3.1.1
pysam==0.22.0
pytest==7.1.3
pytest-azurepipelines==1.0.5
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-nunit==1.0.3
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2023.3.post1
readchar==4.0.5
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
RUST==1.3.1
semantic-version==2.10.0
setuptools-rust==1.7.0
six==1.16.0
soupsieve==2.5
tblib==1.7.0
tk==0.1.0
tomli==2.0.1
typing-extensions==4.8.0
urllib3==1.26.17
wcwidth==0.2.8
zeep==4.2.1
zipp==3.17.0
